Reworked implementation of SchedulerBase::process. The previous implementation basically forced gated-polling approach, and simulated exhaustive polling by setting the gating value extremely high.
Extracted this implementation out into SchedulingPolicy class by adding
some functions to the SchedulingPolicy interface:
- void batchStart(PollingInformation& pollingInfo);
- void batchEnd(PollingInformation& pollingInfo);
- bool continueProcessing(PollingInformation& pollingInfo);

batchStart() tells the SchedulingPolicy that we are starting to process
a batch of messages. This is to enable the implementation of time-gated
scheduling policies.

Likewise, batchEnd() tells the SchedulingPolicy that we have completed a
batch of messages (because it told us to stop). This couples with
batchStart() for booking keeping of time-gated or size-gated scheduling
policies.

continueProcessing() now implements the gating policy in the
SchedulingPolicy class. This is a big improvement to the flexibility of
the toolkit with respect to polling policies. We are no longer required
to simulate exhaustive polling.

This change introduced the removal of two previously required functions:
- batchSize()
- emptyRetryCount()

TODO: move unit_test/test implementations of exhaustive polling into
wield namespace for reuse. Implement time-gated, and size-gated polling
policies in the wield namespace for user convenience.