Bug Fix:
There was a very subtle problem with the EventSourceQueueImpl since the changes for bug #2. Namely, the reference count of the message starts at 0. And since we expect boost::intrusive_ptr to cleanup the memory in the context calling try_pop() the memory was leaked.

To fix the problem, we have to increment the reference count to 1 manually in our event source queue :/